Crate schema

Structs

ConnectorContext

Connector meta information, to be used in query execution if necessary.

DatabaseEnumType

Deprecation

FieldRefEnumType

Identifier

InputField

InputObjectType

InputObjectTypeConstraints

ObjectType

OutputField

QueryInfo

Designates a specific top-level operation on a corresponding model.

QuerySchema

The query schema. Defines which operations (query/mutations) are possible on a database, based on the (internal) data model.

StringEnumType

Enums

EnumType

IdentifierType

Enum used to represent unique schema type names. It helps deferring the allocation + formatting of strings during the initialization of the schema, which proved to be very costly on large schemas.

InputType

ObjectTag

Object tags help differentiating objects during parsing / raw input data processing, especially if complex object unions are present.

OutputType

QueryTag

A QueryTag designates a top level query possible with Prisma.

ScalarType

Traits

IntoArc

Since we have the invariant that the weak refs that are used throughout the query schema have to be always valid, we use this simple trait to keep the code clutter low.

QuerySchemaRenderer

Trait that should be implemented in order to be able to render a query schema.

Functions

capitalize

scalar_filter_name

Compute the name of a scalar filter input.

Type Definitions

EnumTypeRef

EnumTypeWeakRef

InputFieldRef

InputObjectTypeStrongRef

InputObjectTypeWeakRef

ObjectTypeStrongRef

ObjectTypeWeakRef

OutputFieldRef

OutputTypeRef

QuerySchemaRef